Move scaler to adjustments module
