+++ src/adjustments.py
@@ -1,5 +1,6 @@
 """Synthetic batch effects"""
 import numpy as np
+import pandas as pd
 
 class Noise(object):
     def __init__(
@@ -39,6 +40,60 @@ class Noise(object):
             array[too_low] = self.min
             array[too_high] = self.max
             return array
+
+def split_discrete_continuous(df):
+    discrete_types = ['object', 'int']
+    discrete = df.select_dtypes(include=discrete_types)
+    continuous = df.select_dtypes(exclude=discrete_types)
+    return discrete, continuous
+
+class Scaler(object):
+    """Scale or unscale a dataframe from [min, max] <-> [0, 1]
+    """
+    def __init__(self):
+        self.col_min = None
+        self.col_max = None
+
+    def squash(self, df):
+        """Adjust the dataframe to the [0, 1] range.
+
+        Arguments:
+            df {pandas.DataFrame} -- The quantitative dataframe to be
+                squashed.
+
+        Returns:
+            pandas.DataFrame -- The squashed dataframe.
+        """
+        discrete, continuous = split_discrete_continuous(df)
+
+        self.col_min = continuous.min()
+        self.col_max = continuous.max()
+        scaled = (continuous - self.col_min) / (self.col_max - self.col_min)
+
+        return pd.concat([discrete, scaled], axis="columns")
+
+    def unsquash(self, df):
+        """Adjust the dataframe back to the original range.
+
+        Arguments:
+            df {pandas.DataFrame} -- The quantitative dataframe to be
+                expanded.
+
+        Returns:
+            pandas.DataFrame -- The dataframe with each column expanded
+                to its original range.
+        """
+        if self.col_min is None or self.col_max is None:
+            raise Exception(
+                "Error: Scaler.squash() must be run "
+                "before Scaler.unsquash() can be used."
+            )
+        discrete, continuous = split_discrete_continuous(df)
+        scaled = continuous * (self.col_max - self.col_min) + self.col_min
+        if discrete.size == 0:
+            return scaled
+        return pd.concat([discrete, scaled], axis="columns")
+
 
 
 if __name__ == "__main__":
